+++ src/main/java/io/github/apricotfarmer11/mods/tubion/TubionMod.java
@@ -4,27 +4,27 @@ import io.github.apricotfarmer11.mods.tubion.command.TubionCommandManager;
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.TubnetCore;
 import io.github.apricotfarmer11.mods.tubion.config.TubionConfig;
 import io.github.apricotfarmer11.mods.tubion.config.TubionConfigManager;
-import io.github.apricotfarmer11.mods.tubion.core.tubnet.game.GameMode;
+import io.github.apricotfarmer11.mods.tubion.core.tubnet.event.TubnetConnectionEvents;
 import io.github.apricotfarmer11.mods.tubion.event.ChatMessageEvent;
 import io.github.apricotfarmer11.mods.tubion.feature.AutoGGL;
 import io.github.apricotfarmer11.mods.tubion.feature.ChatUtils;
 import io.github.apricotfarmer11.mods.tubion.feature.discord.SDKInitializer;
 import io.github.apricotfarmer11.mods.tubion.feature.discord.TubnetDiscordIntegration;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.loader.api.FabricLoader;
 //#if MC>=11902
 //$$ import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 //#else
 import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
-import net.minecraft.client.RunArgs;
-import net.minecraft.item.ItemStack;
 //#endif
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.UUID;
 
 public class TubionMod implements ClientModInitializer {
     private static TubnetCore tubnetConHandler;
+    public static HashMap<UUID, Integer> playersUsingTubion = new HashMap<>();
     public static final String VERSION = FabricLoader.getInstance().getModContainer("tubion").get().getMetadata().getVersion().getFriendlyString();
 
     public static TubnetDiscordIntegration discordIntegration;
@@ -39,9 +39,6 @@ public class TubionMod implements ClientModInitializer {
         }
         tubnetConHandler = TubnetCore.getInstance();
         SDKInitializer.loadSDK();
-
-        // Load Discord
-        discordIntegration = new TubnetDiscordIntegration();
         //#if MC>=11902
         //$$ ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> {commandManager = new TubionCommandManager(dispatcher);commandManager.loadCommands();});
         //#else
@@ -52,6 +49,20 @@ public class TubionMod implements ClientModInitializer {
         // Load chat utilities
         ChatMessageEvent.EVENT.register(new ChatUtils()::onChat);
         new AutoGGL();
+        TubnetConnectionEvents.CONNECT.register(() -> {
+            this.initDiscord();
+        });
+        TubnetConnectionEvents.DISCONNECT.register(() -> {
+            discordIntegration.destroy();
+            discordIntegration = null;
+        });
+    }
+    public void initDiscord() {
+        if (discordIntegration != null && discordIntegration.discordGameSdk != null && discordIntegration.discordGameSdk.isOpen()) {
+            discordIntegration.destroy();
+            discordIntegration = null;
+        }
+        discordIntegration = new TubnetDiscordIntegration();
     }
 
     public static TubionMod getInstance() {return instance;}