+++ src/main/java/io/github/apricotfarmer11/mods/tubion/core/helper/ChatHelper.java
@@ -5,12 +5,18 @@ import net.minecraft.text.MutableText;
 import net.minecraft.util.Formatting;
 
 public class ChatHelper {
-    public static MutableText getChatPrefix() {
-        final MutableText prefix = TextUtils.literal("Tubion").formatted(Formatting.BOLD, Formatting.RED);
-        return prefix.append(" ").append(TextUtils.literal(">").formatted(Formatting.BLUE).append(TextUtils.literal(" ").formatted(Formatting.RESET)));
-    }
     public static MutableText getChatPrefixWithFeature(MutableText feature) {
         final MutableText prefix = TextUtils.literal("Tubion").formatted(Formatting.BOLD, Formatting.GOLD);
-        return prefix.append(TextUtils.literal(" | ").formatted(Formatting.BLUE).append(feature).append(" ").append(TextUtils.literal(">").formatted(Formatting.BLUE).append(TextUtils.literal(" ").formatted(Formatting.RESET))));
+        return prefix
+                .append(
+                        TextUtils.literal("|")
+                                .formatted(Formatting.BLUE)
+                )
+                .append(feature)
+                .append(" ")
+                .append(TextUtils.literal(">").formatted(Formatting.BLUE))
+                .append(
+                        TextUtils.literal(" ").formatted(Formatting.RESET)
+                );
     }
 }