+++ src/main/java/io/github/apricotfarmer11/mods/tubion/core/tubnet/game/mode/Lobby.java
@@ -2,7 +2,6 @@ package io.github.apricotfarmer11.mods.tubion.core.tubnet.game.mode;
 
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.game.TeamType;
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.game.TubnetGame;
-import net.minecraft.client.MinecraftClient;
 
 public class Lobby implements TubnetGame {
     @Override