+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ### Bugfixes
-* Fix "Discord RPC" from spamming "(Duos)" in game state
-* Prevent a crash occurring when a client does not have Mod Menu installed and attempts to load the settings page via `/tubion`.
+* Fix "AutoGG" from not working in specific games.
+* Discord RPC now does not duplicate
+* 
+
+### New Features
+* New Feature: "Hide Tips"
+* Tubion icon next to usernames